#89 - Moving and adapting easyqueue tests to the new module structure
